refactor : outbox 위치 이동
