+++ src/main/java/com/yeoljeong/tripmate/notification/infrastructure/persistence/outbox/NotificationSendOutboxJpaRepository.java
@@ -1,6 +1,5 @@
-package com.yeoljeong.tripmate.notification.infrastructure.persistence.jpa;
+package com.yeoljeong.tripmate.notification.infrastructure.persistence.outbox;
 
-import com.yeoljeong.tripmate.notification.domain.model.NotificationSendOutbox;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
